Fixed paths and removed lambdas
